+++ apps/desktop/src/app/contrib/controller.test.ts
@@ -1,14 +1,16 @@
-import { afterEach, expect, it, vi } from 'vitest'
+import { afterEach, expect, it, onTestFinished, vi } from 'vitest'
 
-vi.mock('@/contrib/plugins', () => ({ discoverBundledPlugins: vi.fn() }))
-vi.mock('@/contrib/runtime-loader', () => ({ discoverRuntimePlugins: vi.fn() }))
+import { registry } from '@/contrib/registry'
+
+import { registerLayoutPresets } from './layout-presets'
 
 afterEach(() => vi.unstubAllGlobals())
 
-it('keeps the main layout set with onboarding disabled', async () => {
+it('keeps the main layout set with onboarding disabled', () => {
   vi.stubGlobal('hermesDesktop', { guestOnboardingEnabled: false })
-  await import('./controller')
-  const { registry } = await import('@/contrib/registry')
+  // The layout invariant does not need the controller's UI imports or watchers.
+  onTestFinished(registerLayoutPresets())
   const layouts = registry.getArea('layouts').map(preset => preset.id)
   expect(new Set(layouts)).toEqual(new Set(['default', 'focus', 'terminal-deck', 'quad']))
+  expect(layouts).not.toContain('basic')
 })
