+++ apps/desktop/src/app/contrib/layout-presets.ts
@@ -1,0 +1,74 @@
+import { group, split } from '@/components/pane-shell/tree/model'
+import { registry } from '@/contrib/registry'
+import { isOnboardingEnabled } from '@/lib/onboarding-enabled'
+
+// ---------------------------------------------------------------------------
+// Layout presets — CHAT (main) always dominates.
+// ---------------------------------------------------------------------------
+
+// The REAL default: sessions left, chat main, and the right sidebars in column
+// order main | … | review | file-browser (files outermost). Each is its OWN
+// zone. Review collapses to nothing while its pane is hidden (⌘G off).
+//
+// Preview tiles are DYNAMIC panes (like session tiles), so no preset names one:
+// they're registered by watchPreviewTiles as tabs open, and dockPaneBeside lands
+// each one directly beside the file tree wherever that currently lives — so a
+// file double-click still slides a preview open as its own pane next to the
+// tree, never as a tab stacked into the files sidebar.
+export const DEFAULT_TREE = split(
+  'row',
+  [
+    group(['sessions'], { id: 'grp-sessions' }),
+    group(['workspace'], { id: 'grp-main' }),
+    split(
+      'column',
+      [
+        split(
+          'row',
+          [group(['review'], { id: 'grp-review' }), group(['files'], { id: 'grp-files' })],
+          [1, 1.2],
+          'spl-rail'
+        ),
+        group(['terminal'], { id: 'grp-terminal' })
+      ],
+      [1.6, 1],
+      'spl-right'
+    )
+  ],
+  [1, 3.4, 1.25],
+  'spl-root'
+)
+
+const FOCUS_TREE = split('row', [group(['sessions']), group(['workspace', 'files', 'review', 'terminal'])], [1, 4.6])
+
+// Basic starts with sessions and chat so first-run users need not learn
+// terminal, files or review panes before using Hermes.
+const BASIC_TREE = split('row', [group(['sessions']), group(['workspace'])], [1, 4.6])
+
+const TERMINAL_TREE = split(
+  'column',
+  [
+    split('row', [group(['sessions']), group(['workspace']), group(['files', 'review'])], [1, 3.2, 1.2]),
+    group(['terminal'])
+  ],
+  [3, 1]
+)
+
+const QUAD_TREE = split(
+  'column',
+  [
+    split('row', [group(['sessions', 'files']), group(['workspace'])], [1, 3]),
+    split('row', [group(['terminal']), group(['review'])], [1.4, 1])
+  ],
+  [3, 1]
+)
+
+export function registerLayoutPresets() {
+  return registry.registerMany([
+    { id: 'default', area: 'layouts', title: 'Default', order: 0, data: DEFAULT_TREE },
+    ...(isOnboardingEnabled() ? [{ id: 'basic', area: 'layouts', title: 'Basic', order: 5, data: BASIC_TREE }] : []),
+    { id: 'focus', area: 'layouts', title: 'Focus', order: 10, data: FOCUS_TREE },
+    { id: 'terminal-deck', area: 'layouts', title: 'Terminal deck', order: 20, data: TERMINAL_TREE },
+    { id: 'quad', area: 'layouts', title: 'Quad', order: 30, data: QUAD_TREE }
+  ])
+}
